feat: tour OTA price comparison + Square BigInt fix
- Fix Square payment: BigInt(amount) -> plain number (JSON serialization)
- New /api/tour-ota-prices route: reads tour_ota_prices from Supabase,
  falls back to static mid-tier prices when cron hasn't run yet
- New TourOTAComparison component: shows GetYourGuide/Viator/TripAdvisor
  prices vs book-direct price, 6% cheaper than lowest OTA
- Experiences listing page: OTA price badges on tour cards (strikethrough
  OTA price + green direct price when available)
- Experiences book page: tours use TourOTAComparison instead of 3-tier
  picker; dinners keep existing tier picker; add-ons unchanged
- useTourOTAPrices hook for bulk OTA price fetching on listing page

To activate: trigger GET /api/cron/tour-prices with CRON_SECRET header
to seed tour_ota_prices via Tavily scraping Viator/GYG

+++ src/app/api/stripe/create-payment-intent/route.ts
@@ -165,7 +165,7 @@ async function trySquarePayment(
       sourceId,
       idempotencyKey,
       amountMoney: {
-        amount: BigInt(Math.round(amount * 100)),
+        amount: Math.round(amount * 100),
         currency: currency.toUpperCase() as any,
       },
       locationId,
